+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Implement `PartialEq<False>` for `True`, and vice-versa.
+- Implement `PartialEq<bool>` for `True` and `False`.
+- Implement `PartialEq<{True, False}>` for `bool`.
 - Mark crate as `#![no_std]`.
 
 ## 0.1.0